requirements.txt
django~=3.1.3
django-cors-headers
djangorestframework~=3.12.2
drf-yasg
whitenoise~=5.2.0
gunicorn
numpy
sklearn
tests~=0.7
requests~=2.24.0